# Faster CS decomposition for equipartitioned 2x2 unitary matrices.