--- a/src/high-level/high-level.lisp
+++ b/src/high-level/high-level.lisp
@@ -777,6 +777,52 @@ with upper left block with dimension P-by-Q. Returns the intermediate representa
                   (make-matrix :rows (- m q) :cols (- m q) :data v2t)
                   (vector-to-list theta)))))))
 
+(deftype complex-matrix-data (num-entries) `(simple-array (complex double-float) (,num-entries)))
+
+(defun csd-2x2-basic (unitary-matrix-2x2 p q)
+  "Returns the Cosine-Sine decomposition of an equipartitioned UNITARY-MATRIX-2x2. The values of P and Q are assumed to be equal to one and ignored. See the documentation of LAPACK-CSD for details about the returned values."
+  ;; This function is meant to be used within LAPACK-CSD in the base case
+  ;; where the unitary matrix is 2x2. It computes the CS decomposition in
+  ;; Lisp and does it faster (more than three times) and with less memory
+  ;; overhead (conses less than half as much) than the LAPACK routine.
+  (declare (type matrix unitary-matrix-2x2)
+           (values matrix matrix matrix matrix list)
+           (ignorable p q)
+           (optimize (speed 3) (safety 0) (debug 0) (space 0)))
+
+  (let* ((data (the (complex-matrix-data 4)
+                    (magicl::matrix-data unitary-matrix-2x2)))
+         (a1 (aref data 0))
+         (a2 (aref data 1))
+         (a3 (aref data 2))
+         (a4 (aref data 3)))
+
+    (let ((c (abs a1))
+          (u1 (cis (phase a1)))
+          (s (abs a2))
+          (u2 (cis (phase a2))))
+
+      (declare (type (double-float -1.0d0 1.0d0) c s)
+               (type (complex double-float) u1 u2)
+               (dynamic-extent c s u1 u2))
+
+      (let ((v2h (conjugate (/ 1.0d0 (- (* c (conjugate u2) a4)
+                                        (* s (conjugate u1) a3))))))
+
+        (let ((mu1 (make-zero-matrix 1 1))
+              (mu2 (make-zero-matrix 1 1))
+              (mv1h (make-zero-matrix 1 1))
+              (mv2h (make-zero-matrix 1 1)))
+
+          (macrolet ((matrix-1x1-data (matrix)
+                       `(the (complex-matrix-data 1) (matrix-data ,matrix))))
+
+            (setf (aref (matrix-1x1-data mu1) 0) u1
+                  (aref (matrix-1x1-data mu2) 0) u2
+                  (aref (matrix-1x1-data mv1h) 0) #c(1.0d0 0.0d0)
+                  (aref (matrix-1x1-data mv2h) 0) v2h))
+
+          (values mu1 mu2 mv1h mv2h (list (atan s c))))))))
 
 ;;; TODO FIXME
 (defun lisp-zlacpy (UPLO M N A rows-a B rows-b &optional (offx 0) (offy 0))