This implements the CS decomposition in the base case of an equipartitioned 2x2 unitary matrix.

The function `csd-2x2-basic` is significantly faster and less memory-hungry than the call to `lapack-csd` as shown below:

```lisp
MAGICL> (let ((p 1)
              (q 1)
              (u (magicl:random-unitary 2)))

          (sb-ext:gc :full t)

          (time
           (dotimes (i 100000)
             (when (= p q 1)
               (csd-2x2-basic u p q))))

          (sb-ext:gc :full t)
           
          (time
           (dotimes (i 100000)
             (lapack-csd u p q))))
Evaluation took:
  0.144 seconds of real time
  0.143749 seconds of total run time (0.130993 user, 0.012756 system)
  [ Run times consist of 0.003 seconds GC time, and 0.141 seconds non-GC time. ]
  100.00% CPU
  390,120,550 processor cycles
  103,980,336 bytes consed
  
Evaluation took:
  0.507 seconds of real time
  0.506777 seconds of total run time (0.488778 user, 0.017999 system)
  [ Run times consist of 0.014 seconds GC time, and 0.493 seconds non-GC time. ]
  100.00% CPU
  1,376,727,302 processor cycles
  246,394,256 bytes consed
```