@stylewarning 2019-12-20
This hasn't received attention in a while. Shall we merge it @jmbr ?