@jmbr 2019-12-23
> This hasn't received attention in a while. Shall we merge it @jmbr ?

Yes.